requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altgraph==0.17.4
asyncio==4.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
edge-tts==7.2.3
frozenlist==1.7.0
gTTS==2.5.4
h11==0.16.0
idna==3.10
imageio==2.37.0
lazy_loader==0.4
lxml==5.4.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pypandoc==1.15
PySocks==1.7.1
python-docx==1.2.0
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
requests==2.32.4
scikit-image==0.25.2
scipy==1.16.2
selenium==4.34.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tabulate==0.9.0
tifffile==2025.9.9
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
